fix chestkyle, restor original script
