+++ jabgui/src/test/java/org/jabref/gui/maintable/MainTableDataModelTest.java
@@ -28,8 +28,8 @@ import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibEntryPreferences;
 import org.jabref.model.entry.field.Field;
 import org.jabref.model.entry.field.StandardField;
-import org.jabref.model.groups.GroupTreeNode;
 import org.jabref.model.groups.GroupHierarchyType;
+import org.jabref.model.groups.GroupTreeNode;
 import org.jabref.model.groups.WordKeywordGroup;
 import org.jabref.model.search.SearchDisplayMode;
 import org.jabref.model.search.query.SearchQuery;
