+++ jablib/src/main/java/org/jabref/model/entry/AuthorList.java
@@ -1,11 +1,8 @@
 package org.jabref.model.entry;
 
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-import java.util.WeakHashMap;
 import java.util.function.Function;
 import java.util.stream.Collector;
 import java.util.stream.Collectors;
